requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
google-auth==2.47.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
idna==3.11
oauthlib==3.3.1
pip==24.3.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
urllib3==2.6.3
pandas
openpyxl